Make GenerateSelfSigned configurable
GenerateSelfSigned was only able to generate an ECDSA certificate with
the P-256 curve. The implementation is now moved into a new method,
GenerateSelfSignedByTypeAndCurve, to support any signing method and curve
that we support. This support is enabled by
dc5ba12df84a0c90b4eb313f6687ccfeccafc827, since namedCurveKeyPair now
implements crypto.Signer we can now use those or the result of
rsa.GenerateKey() when calling x509.CreateCertificate().

The previous commits of cf1fa7deb8b28f2a6a585892cecad5fa85835d7e and
e3748d784af82b0d846514e50a2969e2c5614315 are necessary because the
function signature of GenerateSelfSignedByTypeAndCurve() needs them.
I would have preferred to let them take a clientCertificateType and
namedCurve directly but since those are unexported only the dtls
package itself can create it.

GenerateSelfSigned() is updated to be just a shim for calling
GenerateSelfSignedByTypeAndCurve() for an ECDSA pair with the P-256
curve as not to break the existing API and expectations on that method.

The examples are updated to use the GenerateSelfSignedByTypeAndCurve()
instead as to expose people to this method since it gives them greater
control over what gets generated.